requirements.txt
streamlit
yfinance
plotly
pandas
numpy
python-dateutil
